+++ jack/src/main/scala/chisel/ResetPipeline.scala
@@ -1,0 +1,5 @@
+// See LICENSE for license details
+package barstools.jack.chisel
+
+import chisel3.{Module, Clock}
+
